proj/requirements.txt
amply==0.1.4
anyio==3.2.1
appdirs==1.4.4
appnope==0.1.2
argon2-cffi==20.1.0
astroid==2.6.0
async-generator==1.10
attrs==21.2.0
Babel==2.9.1
backcall==0.2.0
black==21.6b0
bleach==3.3.0
Brotli==1.0.9
certifi==2021.5.30
cffi==1.14.5
chardet==4.0.0
click==8.0.1
dash==1.20.0
dash-core-components==1.16.0
dash-html-components==1.1.3
dash-renderer==1.9.1
dash-table==4.11.3
decorator==5.0.9
defusedxml==0.7.1
docutils==0.17.1
entrypoints==0.3
Flask==2.0.1
Flask-Compress==1.10.1
Flask-SQLAlchemy==2.5.1
future==0.18.2
greenlet==1.1.0
gunicorn==20.1.0
idna==2.10
ipykernel==5.5.5
ipython==7.24.1
ipython-genutils==0.2.0
ipywidgets==7.6.3
isort==5.9.1
itsdangerous==2.0.1
jedi==0.18.0
Jinja2==3.0.1
json5==0.9.6
jsonschema==3.2.0
jupyter-client==6.1.12
jupyter-console==6.4.0
jupyter-core==4.7.1
jupyter-server==1.9.0
jupyterlab==3.0.16
jupyterlab-pygments==0.1.2
jupyterlab-server==2.6.0
jupyterlab-widgets==1.0.0
lazy-object-proxy==1.6.0
MarkupSafe==2.0.1
matplotlib-inline==0.1.2
mccabe==0.6.1
mistune==0.8.4
mypy-extensions==0.4.3
nbclassic==0.3.1
nbclient==0.5.3
nbconvert==6.1.0
nbformat==5.1.3
nest-asyncio==1.5.1
notebook==6.4.0
numpy==1.21.0
packaging==20.9
pandas==1.2.5
pandocfilters==1.4.3
parso==0.8.2
pathspec==0.8.1
pexpect==4.8.0
pickleshare==0.7.5
plotly==5.0.0
prometheus-client==0.11.0
prompt-toolkit==3.0.19
psycopg2==2.9.1
ptyprocess==0.7.0
PuLP==2.4
pycparser==2.20
Pygments==2.9.0
pylint==3.0.0a3
pylint-flask==0.6
pylint-flask-sqlalchemy==0.2.0
pylint-plugin-utils==0.6
pyparsing==2.4.7
pyrsistent==0.17.3
python-dateutil==2.8.1
python-dotenv==0.18.0
pytz==2021.1
pyzmq==22.1.0
qtconsole==5.1.0
QtPy==1.9.0
regex==2021.4.4
requests==2.25.1
requests-unixsocket==0.2.0
Send2Trash==1.7.1
six==1.16.0
sniffio==1.2.0
SQLAlchemy==1.4.19
tenacity==7.0.0
terminado==0.10.1
testpath==0.5.0
toml==0.10.2
tornado==6.1
traitlets==5.0.5
urllib3==1.26.5
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==1.1.0
Werkzeug==2.0.1
widgetsnbextension==3.5.1
wrapt==1.12.1
